[backend] refacto to engine-data-converter (#9122)
